refactor: change paths for tests
