+++ src/components/About/About.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it } from 'vitest';
-import About from '../../components/About/About';
+import About from './About';
 
 describe('About', () => {
   it('About component', () => {
