+++ src/components/Header/Header.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { describe, it } from 'vitest';
-import Header from '../../components/Header/Header';
+import Header from './Header';
 
 describe('Header', () => {
   it('renders header element', () => {
