+++ src/components/Home/ListItems/ListItems.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it } from 'vitest';
-import ListItems from '../../components/Home/ListItems/ListItems';
+import ListItems from './ListItems';
 
 describe('List Items', () => {
   it('Making a list of items', () => {
